+++ features/admin.sms-providers.v1/constants/sms-provider-constants.ts
@@ -69,6 +69,11 @@ export class SMSProviderConstants {
             key: AuthType.API_KEY,
             text: "externalApiAuthentication:fields.authentication.types.apiKey.name",
             value: AuthType.API_KEY
+        },
+        {
+            key: AuthType.PASSWORD_CREDENTIAL,
+            text: "externalApiAuthentication:fields.authentication.types.passwordCredential.name",
+            value: AuthType.PASSWORD_CREDENTIAL
         }
     ];